Move SIMSOPT-compatible VMEC++ wrapper to main Python sources
Instead of keeping it hidden among the C++ sources and re-exporting
it from the main Python sources.
