+++ tests/test_simsopt_compat.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: 2024-present Proxima Fusion GmbH <info@proximafusion.com>
 #
 # SPDX-License-Identifier: MIT
-"""Basic execution tests for SIMSOPT-compatible Python wrapper of VMEC++."""
+"""Tests for VMEC++'s'SIMSOPT compatibility layer."""
 
 from pathlib import Path
 
@@ -9,14 +9,14 @@ import netCDF4
 import numpy as np
 import pytest
 
-from vmecpp.cpp.vmecpp.vmec.pybind11 import simsopt_vmecpp
+from vmecpp import simsopt_compat
 
 # We don't want to install tests and test data as part of the package,
 # but scikit-build-core + hatchling does not support editable installs,
 # so the tests live in the sources but the vmecpp module lives in site_packages.
 # Therefore, in order to find the test data we use the relative path to this file.
 # I'm very open to alternative solutions :)
-REPO_ROOT = Path(__file__).parent.parent.parent.parent.parent.parent
+REPO_ROOT = Path(__file__).parent.parent
 TEST_DATA_DIR = REPO_ROOT / "src" / "vmecpp" / "cpp" / "vmecpp" / "test_data"
 
 
@@ -26,8 +26,8 @@ def json_input_filepath() -> Path:
 
 
 @pytest.fixture
-def vmec(json_input_filepath) -> simsopt_vmecpp.Vmec:
-    return simsopt_vmecpp.Vmec(json_input_filepath)
+def vmec(json_input_filepath) -> simsopt_compat.Vmec:
+    return simsopt_compat.Vmec(json_input_filepath)
 
 
 @pytest.fixture
@@ -111,7 +111,7 @@ def test_wout_attributes_shape(vmec, attribute_name, mnmax_size_name):
 
 def test_changing_boundary():
     # this test only makes sense for a circular tokamak setup
-    vmec = simsopt_vmecpp.Vmec(TEST_DATA_DIR / "circular_tokamak.json")
+    vmec = simsopt_compat.Vmec(TEST_DATA_DIR / "circular_tokamak.json")
     original_rc00 = vmec.boundary.get_rc(0, 0)
     vmec.run()
     assert vmec.wout is not None
